+++ CHANGELOG.md
@@ -4,11 +4,12 @@ Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<ma
 ## [v2.1.0]
 ### Added
 * #49 Add support for Flightradar24 API
+* Explicitly disable USB SD passthrough in `settings.toml`
 
 ### Changed
-* Bump supported CircuitPython to v10.2 only
-* Bump vendored libraries to v10.2 compatible `*.mpy` files
-* Bump vendored `mpy-cross` binaries to CircuitPython v10.2
+* Bump supported CircuitPython to v10.3 only
+* Bump vendored libraries to v10.3 compatible `*.mpy` files
+* Bump vendored `mpy-cross` binaries to CircuitPython v10.3
 * #47 Update OpenSky API for new authentication flow
 * Migrate from `settings.py` to `settings.toml` for secrets management